requirements.txt
minio
flask
paho-mqtt
requests